requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-crispy-forms==1.10.0
idna==2.10
pkg-resources==0.0.0
plaid-python==7.2.0
pytz==2020.4
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.26.2
